Refactor: Split big script into a small lib
